chore: rename cnn_builder to pytorch_cnn_builder
Also extract padded_features and overlapped_features implementations
